+++ exercises/raindrops/src/lib.rs
@@ -1,3 +1,3 @@
-pub fn raindrops(n: usize) -> String {
+pub fn raindrops(n: u32) -> String {
     unimplemented!("what sound does Raindrop #{} make?", n)
 }
